// src/app/shared/enums/filter-type.enum.ts
export enum FilterTypeEnum {
  TEXT = 'text',
  SELECT = 'select',
  MULTISELECT = 'multiselect',
  DATE_RANGE = 'date-range',
  AGENT_SELECTOR = 'agent-selector',
  CLIENT_SELECTOR = 'client-selector',
  CURRENCY = 'currency',
  PAYMENT_ENTITY = 'payment-entity',
  POLICIES_NEARING_EXPIRATION = 'policies-nearing-expiration',
}